Improve indentation behavior.
Indenting a line in Dyalog mode now behaves properly, i.e if point is before
the current level of indentation it moves to the beginning of indentation, if it
is after, it stays put, even if the contents of the line move left or right.

This also fixes a bug in Emacs 24.4 which would reset the indentation due to the
save-excursion, causing indentation to fail.