Fix various jit-lock bugs.
1) syntax-ppss was sometimes unreliable when called from jit-lock. Now we detect
this and fall back on scan-lists instead, since that seems to work.

2) dyalog-fontify locals would fontify outside of the supplied region for long
functions, since it always moved to end-of-header, even when the function header
was located before the start of the region to jit-lock. Fixing this improved
performance.

3) The improvements to defun-navigation to include d-funs caused errors in
dyalog-tradfn-info, since that assumed that dyalog-previous-defun always ended
up at the top of the current tradfn.