Fix infinite loop when indentation decreased.
Since the end position wasn't a marker, and re-indenting lines can remove
characters, the end position might never be reached (if indenting the whole
buffer), causing an infinite loop.

Sadly, this affects performance rather severly, making indenting 30% slower.
This can likely be improved for the common case of small indentation changes by
not calling insert and delete-char for every line with a label.