Always return nil from beginning-of-defun-function.
It would be nice if we return t or nil here, since that mimics how things should
work, but then beginning-of-defun will move to the beginning of the line, which
breaks moving to the start of dfns, so we always return nil.

Hopefully this doesn't break any general minor modes, we can always call
dyalog-next-defun and dyalog-previous-defun ourselves if we need information
on if we found a defun or not.